+++ src/main/java/com/yeoljeong/tripmate/domain/enums/Country.java
@@ -1,0 +1,5 @@
+package com.yeoljeong.tripmate.domain.enums;
+
+public enum Country {
+  JP
+}
