+++ src/main/java/com/yeoljeong/tripmate/domain/enums/ParticipantStatus.java
@@ -1,4 +1,4 @@
-package com.yeoljeong.tripmate.domain.participation.enums;
+package com.yeoljeong.tripmate.domain.enums;
 
 public enum ParticipantStatus {
   PENDING, APPROVAL, REJECTED
