+++ src/main/java/com/yeoljeong/tripmate/domain/enums/PlanCreationType.java
@@ -1,0 +1,5 @@
+package com.yeoljeong.tripmate.domain.enums;
+
+public enum PlanCreationType {
+  CUSTOM
+}
