+++ src/main/java/com/yeoljeong/tripmate/domain/enums/RecruitStatus.java
@@ -1,0 +1,6 @@
+package com.yeoljeong.tripmate.domain.enums;
+
+public enum RecruitStatus {
+  OPEN,
+  CLOSE
+}
