--- src/main/java/com/yeoljeong/tripmate/domain/participation/enums/ParticipantRole.java
@@ -1,5 +1,0 @@
-package com.yeoljeong.tripmate.domain.participation.enums;
-
-public enum ParticipantRole {
-  HOST, GUEST
-}
