--- src/main/java/com/yeoljeong/tripmate/domain/plan/enums/PlanCreationType.java
@@ -1,5 +1,0 @@
-package com.yeoljeong.tripmate.domain.plan.enums;
-
-public enum PlanCreationType {
-  CUSTOM
-}
