--- src/main/java/com/yeoljeong/tripmate/domain/plan/enums/RecruitStatus.java
@@ -1,6 +1,0 @@
-package com.yeoljeong.tripmate.domain.plan.enums;
-
-public enum RecruitStatus {
-  OPEN,
-  CLOSE
-}
